What a mess

+++ FastNoise/Noises/Value/ValueFractalFBM.cs
@@ -13,18 +13,18 @@ namespace FastNoise.Noises
             _interpolator = interpolator;
             _noiseSettings = noiseSettings;
         }
-        public double GetNoise(Vector2 vec)
+        public float GetNoise(Vector2 vec)
         {
             throw new NotImplementedException();
         }
 
-        public double GetNoise(Vector3 vec)
+        public float GetNoise(Vector3 vec)
         {
             var seed = _noiseSettings.Seed;
 
-            double sum = new ValueNoise(_interpolator, _noiseSettings).GetNoise(vec);
+            float sum = new ValueNoise(_interpolator, _noiseSettings).GetNoise(vec);
 
-            double amp = 1;
+            float amp = 1;
 
             var originalSeed = _noiseSettings.Seed;
